+++ tests/integrations/test_pyramid.py
@@ -8,7 +8,7 @@ import pytest
 
 import svcs
 
-from tests.fake_factories import int_factory
+from tests.fake_factories import int_factory, nop
 from tests.ifaces import AnotherService, Service
 
 
@@ -21,20 +21,20 @@ except ImportError:
     pytest.skip("Pyramid not installed", allow_module_level=True)
 
 
-def make_config(view=None):
+def make_config():
     config = Configurator(settings={})
     svcs.pyramid.init(config)
 
-    if view is not None:
-        config.add_route(view, "/")
+    config.add_route("tl_view", "/tl")
+    config.add_route("health_view", "/health")
+
     config.scan()
 
     return config
 
 
-def make_test_app(view=None):
-    config = make_config(view)
-
+@pytest.fixture(name="tapp")
+def _tapp(config):
     return webtest.TestApp(config.make_wsgi_app())
 
 
@@ -93,12 +93,10 @@ def tl_view(request):
     return {"svc": svc}
 
 
-def test_thread_locals():
+def test_thread_locals(tapp):
     """
     Thread locals are available in views.
     """
-    tapp = make_test_app("tl_view")
-
     closed = False
 
     def closing_factory():
@@ -111,5 +109,24 @@ def test_thread_locals():
     svcs.pyramid.register_value(tapp.app, AnotherService, 23)
     svcs.pyramid.register_factory(tapp.app, float, closing_factory)
 
-    assert {"svc": 42} == tapp.get("/").json
+    assert {"svc": 42} == tapp.get("/tl").json
     assert closed
+
+
+@view_config(route_name="health_view", renderer="json")
+def health_view(request):
+    assert (
+        svcs.pyramid.get_pings(request)
+        == svcs.pyramid.get_pings()
+        == svcs.pyramid.svcs_from(request).get_pings()
+    )
+    return {"num": len(svcs.pyramid.svcs_from(request).get_pings())}
+
+
+def test_get_pings(tapp):
+    """
+    get_pings() returns service pings with and without passing a request.
+    """
+    svcs.pyramid.get_registry(tapp.app).register_value(Service, 42, ping=nop)
+
+    assert {"num": 1} == tapp.get("/health").json